+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 0.4.0-dev
+Version 0.4.0
 - PhptRunner now reports parse error as a failure
 - added TestsRunner
 - common ini settings can now be set for PhpRunner